Amélioration visuel home

+++ projet-vente-en-ligne/src/Service/ProduitService.php
@@ -48,8 +48,26 @@ class ProduitService
 
     public function supprimerProduit(int $id): void
     {
+        $references = $this->produitRepository->verifierReferencesCommande($id);
+    
+        if ($references > 0) {
+            throw new \Exception("Impossible de supprimer le produit. Il est référencé dans des commandes.");
+        }
+    
         $this->produitRepository->delete($id);
     }
+
+    public function recupererProduitsParCategorie(): array
+{
+    return $this->produitRepository->findProduitsParCategorie();
+}
+
+public function recupererProduitsSansCategorie(): array
+{
+    return $this->produitRepository->findProduitsSansCategorie();
+}
+
+    
 
     /**
      * Récupérer des produits pour affichage (e.g., en promotion ou les plus populaires).
